requirements.txt
asgiref==3.9.0
async-timeout==5.0.1
attrs==25.3.0
Django==5.2.4
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
psycopg2-binary==2.9.10
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
ruff==0.12.2
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
